+++ src/components/Hero/Hero.stories.tsx
@@ -229,6 +229,46 @@ export const Campaign: Story = {
           subHeaderText={otherSubHeaderTextLong}
         />
       </div>
+      <div>
+        <Heading
+          id="campaign-hero-long-text"
+          text="Campaign Hero with backdrop background color"
+        />
+        <Hero
+          backdropBackgroundColor="section.research.primary"
+          heroType="campaign"
+          heading={
+            <Heading
+              level="one"
+              id="campaign-hero-long-text-heading"
+              text="Hero Campaign"
+            />
+          }
+          imageProps={imageProps}
+          subHeaderText={otherSubHeaderTextLong}
+        />
+      </div>
+      <div>
+        <Heading
+          id="campaign-hero-long-text"
+          text="Campaign Hero with separate backdrop and foreground background color"
+        />
+        <Hero
+          backdropBackgroundColor="section.education.primary"
+          backgroundColor="#DC8034"
+          foregroundColor="black"
+          heroType="campaign"
+          heading={
+            <Heading
+              level="one"
+              id="campaign-hero-long-text-heading"
+              text="Hero Campaign"
+            />
+          }
+          imageProps={imageProps}
+          subHeaderText={otherSubHeaderTextLong}
+        />
+      </div>
     </Stack>
   ),
 };